- the "last updated" label now counts the item totals fetch, so it can no longer claim data is fresher than it is - the analytics dashboard moved to its own page, with admin home wrapping it to show the checkout message - building the CSV report moved out of the page into a tested helper - fetching a date range and its comparison period moved into the analytics service
