+++ src/components/Analytics/LowStockTable.tsx
