+++ src/components/Analytics/RankedBarChart.tsx
