+++ src/components/Analytics/ResidentsDetailTable.tsx
