+++ src/pages/Analytics/index.test.tsx
@@ -14,7 +14,7 @@ import {
 import '@testing-library/jest-dom';
 import { describe, test, expect, vi, beforeEach, afterEach } from 'vitest';
 import { MemoryRouter } from 'react-router-dom';
-import AdminHome from './index';
+import Analytics from './index';
 import { UserContext } from '../../components/contexts/UserContext';
 import * as historyService from '../../services/historyService';
 import * as analyticsService from '../../services/analyticsService';
@@ -30,7 +30,11 @@ import {
 } from '../../types/interfaces';
 
 vi.mock('../../services/historyService');
-vi.mock('../../services/analyticsService');
+// Partial: fetchRangeData stays real so it still calls the mocked history service.
+vi.mock('../../services/analyticsService', async (importOriginal) => ({
+  ...(await importOriginal<typeof import('../../services/analyticsService')>()),
+  getCheckoutItemTotals: vi.fn(),
+}));
 vi.mock('../../services/itemsService');
 vi.mock('../../services/residentService');
 vi.mock('../../components/History/CustomDateDialog', () => ({
@@ -158,7 +162,7 @@ const mockInventoryHistory: InventoryTransaction[] = [
   },
 ];
 
-const renderAdminHome = () =>
+const renderAnalytics = () =>
   render(
     <MemoryRouter>
       <UserContext.Provider
@@ -174,12 +178,12 @@ const renderAdminHome = () =>
           setPinVerified: vi.fn(),
         }}
       >
-        <AdminHome />
+        <Analytics onError={vi.fn()} />
       </UserContext.Provider>
     </MemoryRouter>,
   );
 
-describe('AdminHome Component', () => {
+describe('Analytics Page', () => {
   beforeEach(() => {
     vi.clearAllMocks();
     // The page caches its fetches in session storage; each test starts cold.
@@ -218,7 +222,7 @@ describe('AdminHome Component', () => {
   };
 
   test('renders tiles with computed values', async () => {
-    renderAdminHome();
+    renderAnalytics();
 
     const residentsTile = await findTileCard('Residents Served');
     expect(within(residentsTile).getByText('2')).toBeInTheDocument();
@@ -228,7 +232,7 @@ describe('AdminHome Component', () => {
   });
 
   test('totals only InventoryAdd quantity in the Items Added tile', async () => {
-    renderAdminHome();
+    renderAnalytics();
 
     // 24 from the InventoryAdd row; the InventoryReplaceValue row's 5 is excluded.
     const addedTile = await findTileCard('Items Added');
@@ -236,7 +240,7 @@ describe('AdminHome Component', () => {
   });
 
   test('renders Top 10 Items Checked Out and Residents Served by Building panels', async () => {
-    renderAdminHome();
+    renderAnalytics();
 
     const itemsPanel = (
       await screen.findByText('Top 10 Items Checked Out')
@@ -252,7 +256,7 @@ describe('AdminHome Component', () => {
   });
 
   test('shows the at/below-threshold item with its status chip in Low Stock table', async () => {
-    renderAdminHome();
+    renderAnalytics();
 
     const lowStockPanel = (
       await screen.findByText('Low Stock & High Need')
@@ -265,7 +269,7 @@ describe('AdminHome Component', () => {
   });
 
   test('detail table renders and flags the duplicate resident with a Repeat chip', async () => {
-    renderAdminHome();
+    renderAnalytics();
 
     const detailPanel = await findDetailPanel();
     const aliceRows = within(detailPanel).getAllByText('Alice Resident');
@@ -279,7 +283,7 @@ describe('AdminHome Component', () => {
   });
 
   test('toggling "Repeats only" hides the non-duplicate rows', async () => {
-    renderAdminHome();
+    renderAnalytics();
 
     const detailPanel = await findDetailPanel();
     expect(within(detailPanel).getByText('Bob Resident')).toBeInTheDocument();
@@ -296,7 +300,7 @@ describe('AdminHome Component', () => {
   });
 
   test('changing the building filter narrows the detail rows to that building', async () => {
-    renderAdminHome();
+    renderAnalytics();
 
     expect(
       within(await findDetailPanel()).getByText('Bob Resident'),
@@ -321,13 +325,13 @@ describe('AdminHome Component', () => {
   });
 
   test('serves a remount from cache and refetches when Refresh is clicked', async () => {
-    const { unmount } = renderAdminHome();
+    const { unmount } = renderAnalytics();
     await findDetailPanel();
     const callsAfterFirstLoad = vi.mocked(historyService.getCheckoutHistory)
       .mock.calls.length;
     unmount();
 
-    renderAdminHome();
+    renderAnalytics();
     await findDetailPanel();
     expect(historyService.getCheckoutHistory).toHaveBeenCalledTimes(
       callsAfterFirstLoad,
@@ -342,7 +346,7 @@ describe('AdminHome Component', () => {
   });
 
   test('shows an enabled Export CSV button', async () => {
-    renderAdminHome();
+    renderAnalytics();
 
     await findDetailPanel();
     const exportButton = screen.getByRole('button', { name: /export csv/i });
@@ -350,7 +354,7 @@ describe('AdminHome Component', () => {
   });
 
   test('omits the per-day average from the Checkouts tile for a single-day range', async () => {
-    renderAdminHome();
+    renderAnalytics();
 
     const tile = await findTileCard('Checkouts');
     expect(within(tile).getByText('3')).toBeInTheDocument();
@@ -358,7 +362,7 @@ describe('AdminHome Component', () => {
   });
 
   test('shows the per-day average beside the total once the range spans days', async () => {
-    renderAdminHome();
+    renderAnalytics();
     await findDetailPanel();
 
     fireEvent.mouseDown(screen.getByRole('combobox', { name: /date/i }));
@@ -370,7 +374,7 @@ describe('AdminHome Component', () => {
   });
 
   test('charts only InventoryAdd rows in Top Inventory Items Added', async () => {
-    renderAdminHome();
+    renderAnalytics();
 
     const panel = (
       await screen.findByRole('heading', {
@@ -383,7 +387,7 @@ describe('AdminHome Component', () => {
   });
 
   test('offers a current-inventory export in the split-button menu', async () => {
-    renderAdminHome();
+    renderAnalytics();
 
     await findDetailPanel();
     fireEvent.click(
@@ -426,7 +430,7 @@ describe('AdminHome Component', () => {
       },
     ]);
 
-    renderAdminHome();
+    renderAnalytics();
 
     const leastPanel = (
       await screen.findByText('Least Checked Out Items')
@@ -444,7 +448,7 @@ describe('AdminHome Component', () => {
   test('empties Top 10 Items Checked Out but still lists the catalog as least checked out when nothing moved', async () => {
     vi.spyOn(analyticsService, 'getCheckoutItemTotals').mockResolvedValue([]);
 
-    renderAdminHome();
+    renderAnalytics();
 
     const topPanel = (
       await screen.findByText('Top 10 Items Checked Out')
@@ -464,7 +468,7 @@ describe('AdminHome Component', () => {
   });
 
   test('detail table shows the per-resident visit count in the # Visits column', async () => {
-    renderAdminHome();
+    renderAnalytics();
 
     const detailPanel = await findDetailPanel();
     const aliceRows = within(detailPanel).getAllByText('Alice Resident');
@@ -483,7 +487,7 @@ describe('AdminHome Component', () => {
   });
 
   test('shows the per-building visit count as a caption distinct from the unique-resident bar value', async () => {
-    renderAdminHome();
+    renderAnalytics();
 
     const buildingsPanel = (
       await screen.findByText('Residents Served by Building')
